requirements.txt
Django==1.11.24
django-cleanup==4.0.0
django-forms-bootstrap==3.1.0
django-taggit==1.2.0
Pillow==7.1.1
pytz==2019.3
youtube-dl==2020.3.8
